Issue 363: Nodejs publish
